fix: Improved URL parsing to support query params
Improved URL parsing by parsing the URL with urlparse.
This resolves the issue where the url was not correctly parsed when
url query parameters were applied.

A test was added to verify the functionality.

The script is renamed to loomdl as the `-` in the name is not supported for
importing the script in the test file. The readme has been adjusted accordingly.

Ruff was used for linting and formatting.

Dependencies for development are managed using uv. However the project should
still run with a basic python installation.

Fixes #1